Update api/server/main.ts to use collections/index.ts
Just prior to Step 4.16, "api/server/collections/index.ts" was added to make it simpler to import the Chats and Messages collections.  Was it intended to use that in api/server/main.ts?  This PR submits that change.  I am just getting started with Typescript, so I hope it is correct that we do not need to explicitly list the index.ts file.